Move default resolver to resolvers folder, add a makeKey method
